Promote `drivers` module out of `core`

+++ norma/core/types.py
@@ -28,7 +28,7 @@ import aiosql.types
 import typic
 
 if TYPE_CHECKING:
-    from . import drivers
+    from norma import drivers
 
 
 ModelT = TypeVar("ModelT")
@@ -52,15 +52,15 @@ class ConnectorProtocol(Protocol[_CtxT, _NoReturnT, _ConnT]):
     def initialize(self) -> _NoReturnT:
         ...
 
-    def connection(self, *, timeout: int = 10, connection: _ConnT = None) -> _CtxT:
+    def connection(self, *, timeout: float = 10, connection: _ConnT = None) -> _CtxT:
         ...
 
     def transaction(
-        self, *, timeout: int = 10, connection: _ConnT = None, rollback: bool = False
+        self, *, timeout: float = 10, connection: _ConnT = None, rollback: bool = False
     ) -> _CtxT:
         ...
 
-    def close(self, timeout: int = 10) -> _NoReturnT:
+    def close(self, timeout: float = 10) -> _NoReturnT:
         ...
 
     @classmethod
